+++ BE/src/main/java/ifive/idrop/auth/filter/VerifyUserFilter.java
@@ -1,7 +1,7 @@
 package ifive.idrop.auth.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import ifive.idrop.dto.request.LoginRequest;
+import ifive.idrop.auth.dto.LoginRequest;
 import ifive.idrop.entity.enums.Role;
 import ifive.idrop.common.exception.CommonException;
 import ifive.idrop.common.dto.ErrorResponse;
