+++ BE/src/main/java/ifive/idrop/driver/controller/DriverController.java
@@ -5,9 +5,8 @@ import ifive.idrop.auth.resolver.Login;
 import ifive.idrop.common.dto.BaseResponse;
 import ifive.idrop.common.exception.CommonException;
 import ifive.idrop.common.exception.ErrorCode;
-import ifive.idrop.dto.request.SubscribeCheckRequest;
+import ifive.idrop.driver.dto.*;
 import ifive.idrop.dto.response.*;
-import ifive.idrop.dto.request.DriverInformation;
 import ifive.idrop.driver.domain.Driver;
 import ifive.idrop.driver.service.DriverService;
 import ifive.idrop.entity.Users;
