+++ BE/src/main/java/ifive/idrop/driver/dto/DriverInformation.java
@@ -1,4 +1,4 @@
-package ifive.idrop.dto.request;
+package ifive.idrop.driver.dto;
 
 import lombok.Getter;
