+++ BE/src/main/java/ifive/idrop/parent/dto/SubscribeRequest.java
@@ -1,4 +1,4 @@
-package ifive.idrop.dto.request;
+package ifive.idrop.parent.dto;
 
 import lombok.Getter;
 import org.json.simple.JSONObject;
