+++ BE/src/main/java/ifive/idrop/parent/service/ParentService.java
@@ -3,11 +3,11 @@ package ifive.idrop.parent.service;
 
 import ifive.idrop.driver.domain.Driver;
 import ifive.idrop.common.dto.BaseResponse;
-import ifive.idrop.dto.request.SubscribeRequest;
+import ifive.idrop.parent.dto.SubscribeRequest;
 import ifive.idrop.dto.response.CurrentPickUpResponse;
 
-import ifive.idrop.dto.response.PickUpHistoryResponse;
-import ifive.idrop.dto.response.ParentSubscribeInfoResponse;
+import ifive.idrop.parent.dto.PickUpHistoryResponse;
+import ifive.idrop.parent.dto.ParentSubscribeInfoResponse;
 import ifive.idrop.entity.*;
 import ifive.idrop.entity.enums.PickUpStatus;
 import ifive.idrop.common.exception.CommonException;
